apps/gq/src/app/process-case-view/tabs/single-quotes-tab/delete-custom-view-modal/delete-custom-view-modal.component.ts
import { Component } from '@angular/core';
import { MatDialogRef } from '@angular/material/dialog';

@Component({
  selector: 'gq-delete-custom-view-modal',
  templateUrl: './delete-custom-view-modal.component.html',
})
export class DeleteCustomViewModalComponent {
  constructor(
    private readonly dialogRef: MatDialogRef<DeleteCustomViewModalComponent>
  ) {}

  closeDialog(): void {
    this.dialogRef.close({ delete: false });
  }

  confirm(): void {
    this.dialogRef.close({ delete: true });
  }
}
